requirements.txt
onvif-zeep==0.2.12
zeep==4.2.1
requests==2.31.0
